poll telemetry, python decouple, refactoring
I was off for some time, because I had final year exams :)

In this update, I implemented Telemetry for polls. For now, it's just monitoring which users viewed the poll.

Also I updated the settings.py to use python-decouple package, because it's much more convenient to have .env file instead of environment vars.

This is just the beginning of telemetry classes and functionality, so expect more implementations, improvements and etc.! :)